Autodetect if default Firebase RTDB should be used.
The CLI will now test for a configured Snapshot Debugger DB of
`{project_id}-cdbg`.  Failing that it will then test the default
Firebase RTDB instance to see if it has been configured for use by the
Snapshot Debugger. This removes the need for the `--use-default-rtdb`
argument from all commands other than the `init` command.

The `--database-url` override option remains when needed to provide
a specific URL to use.