feat: Enable general location support in the CLI (#143)
This change allows the user to specify an arbitrary region via the `--location` option to the `init` command. The `--location` option previously existed, however it only allowed the default value of `us-central1`. While the URL of databases in different regions will have different naming conventions, the CLI already made no assumptions in this regard and no extra changes were required to find the DB URL to use. The changes were the following:

- The `init` command no longer enforces the default of `us-central1` for the `--location` argument.
- Some enhancements to checks and user messages on errors were added for the `init` command when `--location` is used.
- The README was updated with documentation regarding regions

fixes #134
fixes #141

---------

Co-authored-by: James McTavish <mctavish@google.com>